feat: methods keys and elements in all KeySet, typings, and allow for typed KeySetAll and KeySetNone
adding `keys` and `elements`, and stricter typings for remove and intersect methods

BREAKING CHANGE: KeySetAll and KeySetNone will have now `keys` and `elements` methods (both the
same) which return an EmptyArray of the type (Key by default). This is a different behaviour more
than an addition, since before keyset type could be relied upon having or not those getters.