feat: keySetAll and KeySetNone with key type
Both KeySetAll and KeySetNone classes now use a generic for the type of the elements of the universe
they represent, same as Some and AllExceptSome

BREAKING CHANGE: While KeySetAll and KeySetNone were before non typed, now they are. Most of the
time no adaptation is needed but in some cases when using TypeScript it may be needed to specify the
type of the universe when creating a desired All or None key sets.